Created models folder
